Sat morning edits
Added svg repo in so we can use ready use vector images
‘—
Adjusted footer content and contact us… need to make it display
properly in contact us.

Yay we’re almost done, one more week i think we’ll launch it.

I will also create a to do list on the root folder to help with the
things that needs to be done. As soon as we finish something from the
list, we can scratch it. If we find something that needs to be done, we
can add it.